loggermanager_azure_appinsights/pyproject.toml
# loggermanager_azure_appinsights/pyproject.toml
[tool.poetry]
name = "loggermanager_azure_appinsights"
version = "0.1.0"
description = "Azure Application Insights plugin for loggermanager."
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
loggermanager = ">=0.1.0"
opentelemetry-api = "1.27.0"
opentelemetry-sdk = "1.27.0"
opentelemetry-exporter-azure-monitor = "1.0.0"

[tool.poetry.plugins."loggermanager.providers"]
azure_appinsights = "loggermanager_azure_appinsights.azure_appinsights_logger:AzureAppInsightsLogger"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
